+++ HISTORY.md
@@ -1,10 +1,12 @@
 # History
 
-## SenkoWSH v3.7.0b
+## SenkoWSH v3.7.0
 ### 更新点
+- `SFile` に圧縮と展開を行う `compress`, `extract` を追加
 - `SFile` にハッシュ値を取得する `getHashCode` を追加
 - `SFile` に `*:` という2文字のパスで `C` 以外を設定した場合に正しいパスを設定できない不具合を修正
 - `SFile` の `getParent`, `getParentFile` でURLを指定した場合に、最後に必ず`\`を付けて返すように動作変更
+- `System` に環境変数を取得する `getEnvironmentString` を追加
 
 ## SenkoWSH v3.6.0
 ### 更新点